Convert supplementary evidence event to java
